+++ src/main/java/com/example/iotl/dto/order/OrderRequestDto.java
@@ -1,4 +1,4 @@
-package com.example.iotl.dto;
+package com.example.iotl.dto.order;
 
 import com.example.iotl.entity.Order.OrderType;
 import java.math.BigDecimal;
